upload

+++ Model/Portfolio/ImageUploader.php
@@ -196,15 +196,16 @@ class ImageUploader
         $baseTmpImagePath = $this->getFilePath($baseTmpPath, $imageName);
 
         try {
-            $this->coreFileStorageDatabase->copyFile(
-                $baseTmpImagePath,
-                $baseImagePath
-            );
-            
-            $this->mediaDirectory->renameFile(
-                $baseTmpImagePath,
-                $baseImagePath
-            );
+            if ($this->getFileInfo()->isExist($imageName, $this->baseTmpPath)) {
+                $this->coreFileStorageDatabase->copyFile(
+                    $baseTmpImagePath,
+                    $baseImagePath
+                );
+                $this->mediaDirectory->renameFile(
+                    $baseTmpImagePath,
+                    $baseImagePath
+                );
+            }
         } catch (\Exception $e) {
             throw new \Magento\Framework\Exception\LocalizedException(
                 __('Something went wrong while saving the file(s).')
@@ -221,13 +222,13 @@ class ImageUploader
      *
      * @deprecated 101.1.0
      */
-    /*private function getFileInfo()
+    private function getFileInfo()
     {
         if ($this->fileInfo === null) {
             $this->fileInfo = ObjectManager::getInstance()->get(FileInfo::class);
         }
         return $this->fileInfo;
-    }*/
+    }
 
     /**
      * Checking file for save and save it to tmp dir
